// backend/src/db/schemas/personal-secrets.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const PersonalSecretsSchema = z.object({
  personalSecretKeyId: z.string().uuid().default("00000000-0000-0000-0000-000000000000"),
  userId: z.string().uuid(),
  secretType: z.string(),
  secretNameCipher: z.string(),
  secretValueCipher: z.string(),
  version: z.number().default(1).nullable().optional(),
  secretNameIV: z.string(),
  secretValueIV: z.string(),
  secretNameAuthTag: z.string(),
  secretValueAuthTag: z.string(),
  algorithm: z.string().default("aes-256-gcm").nullable().optional(),
  organizationId: z.string().uuid(),
  createdAt: z.date().nullable().optional(),
  updatedAt: z.date().nullable().optional()
});

export type TPersonalSecrets = z.infer<typeof PersonalSecretsSchema>;
export type TPersonalSecretsInsert = Omit<z.input<typeof PersonalSecretsSchema>, TImmutableDBKeys>;
export type TPersonalSecretsUpdate = Partial<Omit<z.input<typeof PersonalSecretsSchema>, TImmutableDBKeys>>;